requirements.txt
torch
torchvision
opencv-python
numpy
pandas
Pillow
